+++ src/components/SignUp/Form/PassWord.tsx
