+++ src/components/SignUp/SignUpForm.tsx
@@ -4,10 +4,10 @@ import palette from '@/lib/styles/palette';
 import { typo } from '@/lib/styles/typo';
 import styled from '@emotion/styled';
 import Button from '../common/atoms/Button';
-import UserName from './UserName';
-import Email from './Email';
-import Password from './PassWord';
-import NickName from './NickName';
+import UserName from './Form/UserName';
+import Email from './Form/Email';
+import Password from './Form/PassWord';
+import NickName from './Form/NickName';
 import React, { useState } from 'react';
 import sw from '@/lib/utils/customSweetAlert';
 import { useSignUpMutation } from '@/api/userApi';
